fix long error

+++ clusterautodownloadplugin/core.py
@@ -174,7 +174,7 @@ class Core(CorePluginBase):
 
     def post_file(self, file_path, file_key):
         auth = get_auth()
-        putpolicy = {'scope':'other-storage:' + file_key,'deadline':str(long(time.time()) * 1000 + 86400000L)}
+        putpolicy = {'scope':'other-storage:' + file_key,'deadline':str(int(time.time()) * 1000 + 86400000)}
         token = auth.uploadtoken(putpolicy)
         param = {'position':'local', 'message':'upload'}
         upload_progress_recorder = UploadProgressRecorder()
